% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index.R
\name{get_index}
\alias{get_index}
\title{Compute survey indices}
\usage{
get_index(data = NULL, new_haul = NULL)
}
\arguments{
\item{data}{list of data provided by the query_data() function (default = NULL)}

\item{new_haul}{dataset of haul locations that include new strata, obtained by running the restrat() function (default = NULL)}
}
\value{
a list that contains original survey indices at the strata (og_indx_st) and subregion (og_indx_reg) scale
and restratified indices at the strata (new_indx_st) and subregion (new_indx_reg) scale.
}
\description{
Function that computes original and re-stratified GOA survey indices (in biomass and numbers).
}
