% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goaRestrat-package.R
\docType{package}
\name{goaRestrat-package}
\alias{goaRestrat}
\alias{goaRestrat-package}
\title{goaRestrat: What the Package Does (One Line, Title Case)}
\description{
Re-stratify GOA survey data.
}
\author{
\strong{Maintainer}: Pete Hulson \email{pete.hulson@noaa.gov}

}
\keyword{internal}
