% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_restrat.R
\name{plot_restrat}
\alias{plot_restrat}
\title{Plot restratification results}
\usage{
plot_restrat(index, species, stock_name)
}
\arguments{
\item{index}{list of computed indices, obtained by running the get_index() function}

\item{species}{species code for which plots will be computed (can also be a vector for stock complexes}

\item{stock_name}{descriptive name for stock, will be included with title and also in plot filename that saved}
}
\value{
plots with time-series and percent difference
}
\description{
Function that plots time-series of orginal indices (with 95\% CIs) and restratified indices (for both biomass and numbers)
and percent difference between original and restratified. Stores plots in 'plots' folder
}
