% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{query_data}
\alias{query_data}
\title{Query data from GOA survey}
\usage{
query_data(species, pull_comp = FALSE)
}
\arguments{
\item{species}{species_codes, i.e., c(10110, 21740)}

\item{pull_comp}{boolean, if TRUE then pull necessary data to compute age/length composition (default = FALSE)}
}
\value{
a list of necessary data sources, dataframes are also written to csv files within the 'data' folder.
}
\description{
Function that connects to AKFIN database (using afscdata package) and pulls catch per-unit-effort (cpue, in numbers and weight),
survey strata (strata), haul information (haul), design-based indices (index), species information (species_names),
and if desired length frequency (lfreq) and age-length specimen (specimen) data from the GAP_PRODUCTS tables
}
\examples{
\dontrun{
query_data(species = c(10110, 21740), pull_comp=TRUE)
}

}
