% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrat.R
\name{restrat}
\alias{restrat}
\title{Re-stratify haul locations}
\usage{
restrat(data = NULL, run_terra = FALSE)
}
\arguments{
\item{data}{list of data provided by the query_data() function (default = NULL)}

\item{run_terra}{boolean, if TRUE then runs the terra::extract() function to re-stratify hauls,
if FALSE, reads in previously run data that is stored in the 'data' folder (default = FALSE). Note: you will need to run this at least once}
}
\value{
a dataset that contains haul information with the original strata as well as the new strata
}
\description{
Function that places historical hauls into the new GOA survey strata. Prior to running this function you will need to
run \code{query_data()} to obtain the list of necessary data (i.e., hauls) as an input to re-stratify the historical hauls.
Note: the meat of this script was obtained from Zack Oyafuso in the GAP program
}
